+++ packages/apptive_grid_core/test/decimal_test.dart
@@ -18,7 +18,6 @@ void main() {
           },
         }
       ],
-      'fieldIds': ['3ftoqhqbct15h5o730uknpvp5'],
       'filter': {},
       'fields': [
         {