+++ packages/apptive_grid_core/lib/model/link/apptive_link_type.dart
@@ -44,6 +44,7 @@ enum ApptiveLinkType {
   remove,
 
   /// Rename a [Space], [Grid], [Form]
+  @Deprecated('Use ApptiveLinkType.patch instead')
   rename,
 
   /// List Shares