+++ packages/apptive_grid_core/test/cross_reference_test.dart
@@ -25,7 +25,6 @@ void main() {
           },
         }
       ],
-      'fieldIds': ['fieldId'],
       'filter': {},
       'fields': [
         {